refactored analysis
