+++ Analysis/pat/nyx/cinema.py
@@ -1,22 +1,29 @@
 #! /usr/bin/env python
-import sys
-sys.path.append(sys.path[0]+"/..")
 
-import argparse, os, csv
+import sys 
+import argparse
+import os
+import csv
+import operator
 
-from pat import file_utilities as futils
-from pat import plot_utilities as putils
-from pat import cinema
-from pat import Job as j
+from pat.utils import file_utilities as futils
+from pat.utils import plot_utilities as putils
+from pat.utils import cinema
+from pat.utils import job as j
 
 
 
-class pat_nyx_cinema(cinema.CinemaWorkflow):
+
+class nyx_cinema(cinema.CinemaWorkflow):
 
 	def prepare_cinema(self):
 		# Open CSV file
-		metrics_csv 	 = self.json_data["project-home"] +  self.json_data['wflow-path'] + "/cbench/" + self.json_data['cbench']['output']['metrics-file'] + ".csv"
-		output_file_name = self.json_data["project-home"] +  self.json_data['wflow-path'] + "/cbench/" + "data.csv"
+		if "metrics-csv" in self.json_data["input"]:
+			metrics_csv = self.json_data["input"]["metrics-csv"]
+		else:
+			metrics_csv = self.json_data["project-home"] +  self.json_data['wflow-path'] + "/cbench/" + self.json_data['cbench']['output']['metrics-file'] + ".csv"
+
+		output_file_name = self.json_data["project-home"] +  self.json_data['wflow-path'] + "/cinema/" + "data.csv"
 
 
 		all = []
@@ -49,13 +56,32 @@ class pat_nyx_cinema(cinema.CinemaWorkflow):
 			futils.write_csv(output_file_name, all)
 
 
+	# Converts a string to an operator
+	def validate(self, operand, relate, result):
+		ops = { "<": operator.lt }
+		return ops[relate]( abs(operand-1.0), result)
+
+	# Process checks found in "cinema-plots" "plotting" "checks"
+	def is_valid(self, pk_ratio):
+		valid = True
+		for check in self.json_data['cinema-plots']['plotting']['checks']:
+			for item in pk_ratio:
+				if not self.validate(item, check['operator'],  check['result']):
+					valid = False 
+					break
+			
+		return valid
+
+				
+
 	def create_plots(self):
 		output_path = self.json_data['project-home'] + self.json_data['wflow-path']
 		output_plot_path = output_path + "/plots"
-		csv_file_path = output_path + "/cbench/" + self.json_data['cbench']['output']['metrics-file'] + ".csv"
 		x_range = self.json_data['cinema-plots']['plotting']['x-range']
 
+
 		for ana in self.json_data['pat']['analysis']:
+
 			plot_title = ana['title']
 			to_plot = []  # all the items to plot
 
@@ -68,17 +94,22 @@ class pat_nyx_cinema(cinema.CinemaWorkflow):
 					k_list  = futils.extract_csv_col(file['path'], ' ', 2)
 					orig_pk = futils.extract_csv_col(file['path'], ' ', 3)
 
+			# Process the other files
 			for file in ana['files']:
 				if (file['name']!="orig"):
-					print (file['path'])
-
 					temp_pk = futils.extract_csv_col(file['path'], ' ', 3)
 					if (temp_pk is not None):
 						pk_ratio = [i / j for i, j in zip(temp_pk, orig_pk)]
 						this_tuple = (pk_ratio, file['name']) #array, name
-						to_plot.append(this_tuple)
 
-			putils.plotScatterGraph(k_list, 'k', 'pk', plot_title, output_plot_path, x_range, to_plot)
+						# Check if passes test
+						if "checks" in self.json_data["cinema-plots"]["plotting"]:
+							if self.is_valid(pk_ratio):
+								to_plot.append(this_tuple)
+						else:
+							to_plot.append(this_tuple)
+
+			putils.plotScatterGraph(k_list, 'k', 'pk-ratio', plot_title, output_plot_path, x_range, to_plot)
 
 
 
@@ -88,7 +119,7 @@ parser.add_argument("--input-file")
 opts = parser.parse_args()
 
 
-# Creat Ciname DB
-cinema = pat_nyx_cinema( opts.input_file )
+# Create Cinema DB
+cinema = nyx_cinema( opts.input_file )
 cinema.create_plots()
 cinema.create_cinema()
