+++ inputs/nyx/test-cinema.json
@@ -1,0 +1,250 @@
+{
+	"project-home": "/projects/exasky/pascal-projects/VizAly-Foresight/",
+	"wflow-path": "testBBB",
+	"foresight-home": "/projects/exasky/pascal-projects/VizAly-Foresight/",
+
+	"input": {
+		"filetype": "NYX",
+		"filename": "/projects/exasky/pascal-projects/VizAly-Foresight/testing/data/z255_32.h5",
+
+		"metrics-csv" : "/projects/exasky/pascal-projects/VizAly-Foresight/testA/metrics_.csv",
+		"analysis-results" : "/projects/exasky/pascal-projects/VizAly-Foresight/testB/",
+
+		"group": "native_fields",
+		"scalars": 
+		[
+			"baryon_density",
+			"dark_matter_density",
+			"temperature",
+			"velocity_x",
+			"velocity_y",
+			"velocity_z"
+		],
+
+		"uncompressed": [
+			{
+				"group": "domain",
+				"attributes": [{
+						"name": "shape",
+						"type": "int",
+						"size": 3
+					},
+					{
+						"name": "size",
+						"type": "double",
+						"size": 3
+					}
+				]
+			},
+
+			{
+				"group": "universe",
+				"attributes": [{
+						"name": "hubble",
+						"type": "double",
+						"size": 1
+					},
+					{
+						"name": "omega_b",
+						"type": "double",
+						"size": 1
+					},
+					{
+						"name": "omega_l",
+						"type": "double",
+						"size": 1
+					},
+					{
+						"name": "omega_m",
+						"type": "double",
+						"size": 1
+					},
+					{
+						"name": "redshift",
+						"type": "double",
+						"size": 1
+					}
+				]
+			}
+		]
+	},
+
+	"compressor-comment": "which compressors to test",
+	"compressors": [{
+			"name": "SZ",
+			"output-prefix": "SZ_",
+			"compressor-params": [{
+					"scalar": [
+						"baryon_density",
+						"dark_matter_density",
+						"temperature"
+					],
+					"abs": 1E-3
+				},
+				{
+					"scalar": [
+						"velocity_x",
+						"velocity_y",
+						"velocity_z"
+					],
+					"pw_rel": 1E-2
+				}
+			]
+		},
+
+		{
+			"name": "BLOSC",
+			"output-prefix": "BLOSC_"
+		},
+
+		{
+			"name": "fpzip",
+			"bits": 24,
+			"output-prefix": "fpzip_24_"
+		},
+
+		{
+			"name": "zfp",
+			"output-prefix": "zfp_abs_.01_",
+			"abs": 1E-2
+		},
+
+		{
+			"name": "fpzip",
+			"bits": 14,
+			"output-prefix": "fpzip_14_"
+		}
+
+	],
+
+	"cbench": {
+		"path": "mpirun $foresight-home$build/CBench",
+
+		"output": {
+			"output-decompressed": true,
+			"output-decompressed-location": "decompressed_files",
+
+			"logfname-comment": "name of output log file",
+			"log-file": "runlog_",
+
+			"metricsfname-comment": "name of file with output",
+			"metrics-file": "metrics_"
+		},
+
+		"metrics": [{
+				"name": "absolute_error"
+			},
+
+			{
+				"name": "relative_error"
+			},
+
+			{
+				"name": "mse"
+			},
+
+			{
+				"name": "psnr"
+			}
+		],
+
+		"evn_path": "scripts/VizAly-CBench.bash.darwin",
+
+		"configuration": {
+			"partition": "galton",
+			"nodes": 1,
+			"ntasks-per-node": 16
+		}
+	},
+
+	"pat": {
+		"evn_path": "scripts/VizAly-CBench.bash.darwin",
+
+		"analysis-tool": {
+			"gimlet-home": "/projects/exasky/gimlet2",
+
+			"analytics": 
+			[
+				{
+					"name": "sim_stats",
+					"type": ["rhob", "rhom", "rhodm", "temp", "velmag", "vz"],
+					"postfix": "_ps3d.txt",
+					"path": "/projects/exasky/gimlet2/apps/sim_stats/sim_stats.ex",
+					"configuration": {
+						"partition": "galton",
+						"nodes": 8,
+						"ntasks-per-node": 8
+					}
+				},
+
+				{
+					"name": "lya_all_axes",
+					"type": ["x", "y", "z"],
+					"postfix": "_flux_ps1d.txt",
+					"path": "/projects/exasky/gimlet2/apps/lya_all_axes/lya_all_axes.ex",
+					"command": "cp /projects/exasky/gimlet2/treecools/TREECOOL_middle .",
+					"configuration": {
+						"partition": "galton",
+						"nodes": 8,
+						"ntasks-per-node": 8
+					}
+				}
+			]
+		},
+
+		"input-files": 
+		[
+            {
+                "output-prefix": "orig",
+                "path": "/projects/exasky/pascal-projects/VizAly-Foresight/testing/data/z255_32.h5"
+            },
+            {
+                "output-prefix": "SZ_",
+                "path": "/projects/exasky/pascal-projects/VizAly-Foresight/testA/SZ___z255_32.h5"
+            },
+            {
+                "output-prefix": "BLOSC_",
+                "path": "/projects/exasky/pascal-projects/VizAly-Foresight/testA/BLOSC___z255_32.h5"
+            },
+            {
+                "output-prefix": "fpzip_24_",
+                "path": "/projects/exasky/pascal-projects/VizAly-Foresight/testA/fpzip_24___z255_32.h5"
+            },
+            {
+                "output-prefix": "zfp_abs_.01_",
+                "path": "/projects/exasky/pascal-projects/VizAly-Foresight/testA/zfp_abs_.01___z255_32.h5"
+            },
+            {
+                "output-prefix": "fpzip_14_",
+                "path": "/projects/exasky/pascal-projects/VizAly-Foresight/testA/fpzip_14___z255_32.h5"
+            }
+        ],
+
+
+		"analysis": []
+	},
+
+
+	"cinema-plots": 
+	{
+		"cinema-db-name": "cinemaDB",
+		"plotting": {
+			"x-range": [0, 11],
+			"checks" :
+			[
+				{
+					"operand" : "pk_ratio",
+					"operator" : "<",
+					"result" : 0.01
+				}
+			]
+		},
+
+		"evn_path": "scripts/VizAly-CBench.bash.darwin",
+		"configuration": {
+			"partition": "galton",
+			"nodes": 1,
+			"ntasks-per-node": 8
+		}
+	}
+}
